requirements.txt
asgiref==3.4.1
Brotli==1.0.9
dj-database-url==1.2.0
Django==3.2.18
djangorestframework==3.14.0
djangorestframework-simplejwt==4.3.0
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
psycopg2==2.9.5
PyJWT==1.7.1
python-dotenv==0.20.0
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.1.1
whitenoise==5.3.0
